requirements.txt
flask
google-generativeai
python-dotenv
pandas
numpy
faiss-cpu
